TP2 Dépôt 04 / Questions 11

+++ TP2_Bieres_BRUN_VERCEIL/src/tp2_bieres_brun_verceil/BouteilleBiere.java
@@ -11,13 +11,13 @@ package tp2_bieres_brun_verceil;
  */
 public class BouteilleBiere {
     String nom;
-    float degreAlcool;
+    double degreAlcool;
     String brasserie;
     boolean ouverte;
     public void lireEtiquette() { 
         System.out.println("Bouteille de " + nom +" (" + degreAlcool + " degres) \nBrasserie : " + brasserie ) ; 
     }
-    public BouteilleBiere(String unNom, float unDegre, String uneBrasserie) {
+    public BouteilleBiere(String unNom, double unDegre, String uneBrasserie) {
         nom = unNom;
         degreAlcool = unDegre;
         brasserie = uneBrasserie;
